+++ src/model/model_pr_action.ts
@@ -1,0 +1,7 @@
+export enum PRAction {
+  READY_FOR_REVIEW = 'Ready for Review',
+  EDITED = 'Edited',
+  CLOSED = 'Closed',
+  OPENED = 'Opened',
+  CONVERTED_TO_DRAFT = 'Converted to Draft',
+}
